requirements.txt
# CodeRefine – Python dependencies

# Web framework & server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Environment variable loading
python-dotenv>=1.0.1

# HTTP client (used for Gemini REST API - avoids protobuf issues on Python 3.14)
httpx>=0.27.0

# AI Providers
groq>=0.9.0        # Groq (LLaMA 3 / Mixtral)
# huggingface_hub  # Optional: only needed if using Hugging Face provider

